wan API: skip invalid charts
If an interface link is down, netdata could return an error like
'Chart is not found: net.ethx'.
Such error creates a JSON with an empty chart for the given interface,
the invalid chart breaks all charts inside the UI.
This commit just avoids to return the chart if it is not available.